pyproject.toml
[project]
name = "disaster-app"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "timm>=0.9.0",
    "peft>=0.6.0",
    "gradio>=4.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "api"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]
